refactor(schema): divide schema-change-utils.ts em diff/grouping/format
Parte da issue #390 (epic #376, quick-wins de fallow). O arquivo tinha
319 linhas e 4 dependentes com responsabilidades bem distintas
(diffing de campo, agrupamento temporal de auditoria, formatação para
UI) — separadas em schema-change-diff.ts, schema-change-grouping.ts e
schema-change-format.ts, cada uma com seu spec de teste dedicado.
Sem mudança de comportamento; schema-utils.ts (primitivas de
versionamento/auditoria) não foi tocado.
